✨ feature : 보유 도서 등록

See merge request s08-webmobile2-sub2/S08P12A209!167

+++ BE/src/main/java/com/seoro/seoro/repository/Book/OwnBookRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.seoro.seoro.domain.entity.Book.OwnBook;
 import com.seoro.seoro.domain.entity.Member.Member;
 
-public interface OwnBookRepository extends JpaRepository<OwnBook, Integer> {
+public interface OwnBookRepository extends JpaRepository<OwnBook, Long> {
 	Long countByMember(Member member);
 	Optional<OwnBook> findByIsbn(String isbn);
 	// List findByBook_Isbn(@Param(value= "isbn") String isbn);
